requirements.txt
openai
gradio
langchain
gpt-index
python-dotenv